startmanager를 projectmanager와 headermanager로 분리
projectmanager: projectname, audiofile 입력
headermanager: kst 파일 header 태그정보 입력
