--- src/container/guest/projectManager/SpeakerList/Detail.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react';
-
-class Detail extends Component {
-    render() {
-        return (
-            <div>
-                detail
-            </div>
-        );
-    }
-}
-
-export default Detail;
